+++ activerecord/CHANGELOG.md
@@ -1,3 +1,9 @@
+*   Raise `ActiveRecord::OrderError` when order dependent finder methods (e.g. `#first`, `#last`) are called
+    without order values, and the model does not have any order columns (`implicit_order_column`, `query_constraints`,
+    or `primary_key`) to fall back on.
+
+    *Joshua Young*
+
 *   Respect `implicit_order_column` in `ActiveRecord::Relation#reverse_order`.
 
     *Joshua Young*